refactor(index): split worktree overlay into cohesive siblings
Decompose linked-worktree overlay handling into discovery, materialization, lifecycle, query-adaptation, and test modules behind a curated compatibility surface.

Preserve dirty overlay behavior, linked worktree identity, base/overlay separation, repo-scoped cleanup, and inline/deferred logical rebuild semantics.

Fixes #920
